Separate out most of the shader program related codes from material.cpp and material.h. #122
